+++ tests/interface/theme_features.py
@@ -8,6 +8,7 @@ SHOW_REVISION_HISTORY = "revision.history"
 HIDE_TOP_NAV = "navigation.top.hide"
 HIDE_TOP_CURSOR_ANIMATION = "navigation.top.cursor_animation.hide"
 HIDE_SIDE_NAV = "navigation.side.hide"
+SHOW_INDEX_SECTIONS = "navigation.side.indexes"
 HIDE_SEARCH_BUTTON = "navigation.top.search_button.hide"
 HIDE_SIDE_TOC = "navigation.side.toc.hide"
 HIDE_LINK_UNDERLINE = "style.links.underline.hide"
